bazar_online_api/pyproject.toml
[project]
name = "bazar-online-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "sqlalchemy>=2.0.38",
    "alembic>=1.14.1",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.2.5",
    "asyncpg>=0.30.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "types-passlib>=1.7.7.20241221",
    "pytest-asyncio>=0.25.3",
    "httpx>=0.28.1",
]

[tool.ruff]
target-version = "py312"
line-length = 119
src = ["app"]

[tool.ruff.format]
line-ending = "lf"
indent-style = "space"
quote-style = "single"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D100", "D104", "D203", "D213", "D401", "D406", "D407", "D413", "COM812", "ISC001", "Q000", "Q003", "TRY003", "EM101"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]

[tool.mypy]
python_version = "3.12"
follow_imports = "normal"
warn_unused_configs = true
show_column_numbers = true
show_error_codes = true
strict = true
plugins = ["pydantic.mypy"]
files = ["app/**/*.py", "tests/**/*.py"]

[tool.pytest.ini_options]
addopts = "--strict-config --cov-report=term-missing --no-cov-on-fail --cov=app"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
  "ignore::DeprecationWarning:passlib.*:",
  "ignore::DeprecationWarning:bcrypt.*:",
  "ignore::pydantic.warnings.PydanticDeprecatedSince20",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if (typing\\.)?TYPE_CHECKING:",
  "@abstractmethod",
  "case _ as unreachable:"
]
